fix: pokemon e pokemondto correcao nos ids

+++ src/main/java/com/dbc/pokesuits/dto/pokemon/PokemonDTO.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 @EqualsAndHashCode(callSuper=false)
 public class PokemonDTO extends PokemonCreateDTO {
 	@ApiModelProperty(name = "Id do Pokemon")
-	private Long idPokemon;
+	private Integer idPokemon;
 }
